Chore: Improve deployment script (#7)

+++ scripts/deploy.sh
@@ -16,12 +16,6 @@ if [[ -z $NETWORK || ! " ${networks[@]} " =~ " ${NETWORK} " ]]; then
   exit 1
 fi
 
-# Validate contract addresses
-if [[ -z $COURT ]]; then
-  echo 'Please make sure a Court address is provided'
-  exit 1
-fi
-
 # Use mainnet network in case of local deployment
 if [[ "$NETWORK" = "rpc" ]]; then
   ENV='mainnet'
@@ -31,11 +25,24 @@ else
   ENV=${NETWORK}
 fi
 
-# Build subgraph manifest for requested variables
-cp subgraph.template.yaml subgraph.yaml
-sed -i -e "s/{{network}}/${ENV}/g" subgraph.yaml
-sed -i -e "s/{{court}}/${COURT}/g" subgraph.yaml
-rm subgraph.yaml-e
+# Validate court address
+if [[ -z $COURT ]]; then
+  if [[ -f subgraph.${NETWORK}.yaml ]]; then
+    # Copy network subgraph manifest
+    echo "Redeploying subgraph using subgraph.${NETWORK}.yml manifest"
+    mv subgraph.${NETWORK}.yaml subgraph.yaml
+  else
+    echo 'Please make sure a Court address is provided'
+    exit 1
+  fi
+else
+  # Build subgraph manifest for requested variables
+  echo "Deploying new subgraph for court address ${COURT} to network ${NETWORK}"
+  cp subgraph.template.yaml subgraph.yaml
+  sed -i -e "s/{{network}}/${ENV}/g" subgraph.yaml
+  sed -i -e "s/{{court}}/${COURT}/g" subgraph.yaml
+  rm subgraph.yaml-e
+fi
 
 # Run codegen
 npm run codegen
@@ -65,3 +72,6 @@ fi
 
 # Deploy subgraph
 graph deploy aragon/aragon-court${SUBGRAPH_EXT} --ipfs ${IPFS_NODE} --node ${GRAPH_NODE}
+
+# Save manifest for custom network
+mv subgraph.yaml subgraph.${NETWORK}.yaml
